server: widen RenderConsumer's emitted-energy accumulators to double
RenderConsumer::total_emitted_energy_ / snapshot_emitted_energy_ were
float, and a long legacy-CPU run (128 rays/batch) charges them hundreds
of thousands of times per render. Once the running total passes float32's
exact-integer ceiling (2^24), every further += rounds in the same
direction, compounding into a systematic 0.05-0.12% drift in the
published LUMICE_RawXyzResult::emitted_energy -- the absolute-EV
denominator. Widen the two accumulators to double; the per-batch addend
(SimData::emitted_energy_) and the published API value stay float, so
this touches no ABI.

Added a unit test that drives 234k batches of ~11258 energy each (the
128-rays/batch, ~30M-ray shape) and asserts the published value against
an exact double reference at <1e-6 relative error. Verified red state by
reverting the two fields to float locally: the same test then fails at
0.0499% drift, matching the measured range. test_ray_allocation_regression.py
and gui_test's export_parity scene (bm4_PSNR=43.59 dB vs a 42.1 dB
threshold) confirm the PSNR/block-mean rulers do not move.

+++ src/server/render.cpp
@@ -250,10 +250,10 @@ float RenderConsumer::ExposureScale() const {
     return 0.0f;
   }
   if (config_.ev_mode_ == RenderConfig::kAbsolute) {
-    if (snapshot_emitted_energy_ <= 0.0f) {
+    if (snapshot_emitted_energy_ <= 0.0) {
       return 0.0f;
     }
-    return config_.intensity_factor_ * kNormScale * total_pix / snapshot_emitted_energy_;
+    return static_cast<float>(config_.intensity_factor_ * kNormScale * total_pix / snapshot_emitted_energy_);
   }
 
   // kRelative: anchor to the SCENE's sky radiance, which is the same number the GUI anchors to.
@@ -1240,7 +1240,7 @@ RawXyzResult RenderConsumer::GetRawXyzResult() const {
                   {},
                   {},
                   effective_pix_,
-                  snapshot_emitted_energy_ };
+                  static_cast<float>(snapshot_emitted_energy_) };
   r.axis_solid_angle_ = AxisSolidAngle();
   return r;
 }
